devops/api/requirement.txt
flask==3.0.2
swagger-ui-py==23.9.23